% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{analysis_exercises_default}
\alias{analysis_exercises_default}
\title{default analysis on exercises level, return r_it, difficulty (P), duration
(dur), number of items}
\usage{
analysis_exercises_default(d)
}
\description{
default analysis on exercises level, return r_it, difficulty (P), duration
(dur), number of items
}
