% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{analysis_items_default}
\alias{analysis_items_default}
\title{d has to be prepared accordingly so that it contains the overall score of the
candidate (score_candidate"). See \code{prepare_items}.}
\usage{
analysis_items_default(d, groups = c("id_answer"))
}
\value{
tibble with id_answer, r_it, P, and n (number of answers)
}
\description{
d has to be prepared accordingly so that it contains the overall score of the
candidate (score_candidate"). See \code{prepare_items}.
}
