% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova}
\alias{anova}
\title{anova exercise}
\usage{
anova(seeds = sample.int(10000, 1), selection = 1:8)
}
\arguments{
\item{seeds}{Seeds for the exercises, can be a single value or a vector for
multiple exercises. Defaults to random seed.}
}
\value{
rqti Entry object
}
\description{
A realistic within ANOVA output is presented based on a real study. 8
questions have to be answered. This is the raw version, producing an rqti
Entry object.
}
\examples{
a <- anova()
}
