% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova_stud}
\alias{anova_stud}
\title{anova test (student version)}
\usage{
anova_stud(seed = 1:20, selection = 1)
}
\arguments{
\item{seed}{Seed for the exercise, can also be a vector to produce several
versions of the exercise. Default is a vector from 1:20.}

\item{selection}{How many exercises should be selected. Has to be smaller
than length of seed.}
}
\value{
rqti test object
}
\description{
A realistic within ANOVA output is presented based on a real study. 8
questions have to be answered. This is the student version, containing 20
versions by default wrapped into a QTI test. One version is selected randomly
by default.
}
\examples{
a <- anova_stud()
}
