% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{bayes}
\alias{bayes}
\title{Bayes exercise}
\usage{
bayes(seed = sample.int(1000))
}
\arguments{
\item{seed}{random seed}
}
\value{
qti object of type Entry
}
\description{
Creates a bayes exercise in qti format
}
