% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{bayes_example}
\alias{bayes_example}
\title{Example data for Bayes Revision}
\usage{
bayes_example(seed = sample.int(1000))
}
\arguments{
\item{seed}{Random seed}
}
\value{
list with elements solution, solution_perc, solution_tree which are
all vectors containg p_hit, p_miss, p_false_alarm, p_correct_rejection,
p_revision, solution has decimal points, solution_perc percentages,
solution_tree for a tree with 10k elements. In addition the list contains
as numerics base_rate, sensitivity and specificity in decimals.
}
\description{
Creates a random fourfould table for bayes exercise.
}
