% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{bayes_feedback}
\alias{bayes_feedback}
\title{generate feedback for bayes exercise}
\usage{
bayes_feedback(example)
}
\arguments{
\item{example}{list from bayes_example}
}
\value{
content for feedback
}
\description{
generate feedback for bayes exercise
}
