% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{bayes_one}
\alias{bayes_one}
\title{Bayes exercise}
\usage{
bayes_one(seed = sample.int(1000, 1))
}
\arguments{
\item{seed}{random seed}
}
\value{
qti object of type Entry
}
\description{
Creates one bayes exercise in qti format
}
