% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{bayes_plot}
\alias{bayes_plot}
\title{create plot for bayes exercise}
\usage{
bayes_plot(example)
}
\arguments{
\item{example}{list from bayes_example}
}
\value{
currently only the captured print since svg does not work in opal?
}
\description{
create plot for bayes exercise
}
