% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{bayes_qdf}
\alias{bayes_qdf}
\title{generate questions for bayes}
\usage{
bayes_qdf(example)
}
\arguments{
\item{example}{bayes_example list}
}
\value{
data frame for df2gap
}
\description{
From bayes_example list, a question data frame is created that can be fed
to df2gap
}
