% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{bayes_story}
\alias{bayes_story}
\title{generate story for bayes exercise}
\usage{
bayes_story(example)
}
\arguments{
\item{example}{list form bayes_example}
}
\value{
currently the story plus the questions
}
\description{
generate story for bayes exercise
}
