% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{df2gap}
\alias{df2gap}
\title{df2gap}
\usage{
df2gap(df)
}
\arguments{
\item{data}{frame that has the same names as NumericGap class, except for
response_identifier, which is taken from the rownames.}
}
\value{
list of NumericGap objects wrapped in html paragraphs.
}
\description{
Takes dataframe of questions and transforms it to NumericGap qti.
}
