% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{df2sc}
\alias{df2sc}
\title{create sc table for df that has different values in cols for each row}
\usage{
df2sc(file)
}
\description{
create sc table for df that has different values in cols for each row
}
