% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{df2sctable}
\alias{df2sctable}
\title{Create sc table for df that has the same values in cols for each row}
\usage{
df2sctable(
  df,
  identifier,
  title = identifier,
  shuffle = T,
  story,
  feedback = "",
  points = nrow(df)/2
)
}
\arguments{
\item{df}{with values rows, cols, rows_id and cols_id}
}
\value{
OneInRowTable item object
}
\description{
Create sc table for df that has the same values in cols for each row
}
\details{
rows is what will be shown in the rows sc table; if you want a
specific order cols should be a factor

Important: To have cols as pure distractors (never correct), cols and cols_id
should be a factor with additional levels!
}
