% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{encodeGraphic}
\alias{encodeGraphic}
\title{encode ggplot graphic as png and return as html string \if{html}{\out{<img>}} in base64}
\usage{
encodeGraphic(g)
}
\arguments{
\item{g}{ggplot2 graphic}
}
\description{
encode ggplot graphic as png and return as html string \if{html}{\out{<img>}} in base64
}
