% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{extract_parameters}
\alias{extract_parameters}
\title{helper function to extract parameters for regression exercise from lm model}
\usage{
extract_parameters(model)
}
\arguments{
\item{model}{lm model object}
}
\value{
list of important parameters, rounded to 3 decimal points
}
\description{
helper function to extract parameters for regression exercise from lm model
}
