% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faktorenanalyse.R
\name{faktorenanalyse}
\alias{faktorenanalyse}
\title{Exercise faktorenanalyse}
\usage{
faktorenanalyse(seeds = sample.int(10000, 1))
}
\arguments{
\item{seeds}{seeds for exercise, can be single value or vector}
}
\description{
Bootstrapped factor analysis based on Thurstone 1941 intelligence data. 4
questions: number of factors for Kaiser criterion and parallel analysis,
calculate communality, calculate variance explained.
}
