% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{get_html_table}
\alias{get_html_table}
\title{Create a html table for lm model}
\usage{
get_html_table(model, sem)
}
\arguments{
\item{sem}{standard error of measurement, which is not available in model and
not generated by tab_model}
}
\description{
writes out sjPlot regression table to table.html
}
