% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{get_questions}
\alias{get_questions}
\title{helper to create questions for regression exercise
q = question, s = solution, p = points, t = tolerance}
\usage{
get_questions(parameters, variables_md)
}
\description{
helper to create questions for regression exercise
q = question, s = solution, p = points, t = tolerance
}
