% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_supplement_paths}
\alias{get_supplement_paths}
\title{Gets all files in inst/supplements of methodenlehre package}
\usage{
get_supplement_paths()
}
\description{
Gets all files in inst/supplements of methodenlehre package
}
