% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ki-bin.R
\name{ki_bin}
\alias{ki_bin}
\title{Exercise ki_bin}
\usage{
ki_bin(seed = sample.int(10000, 1))
}
\value{
: Entry rqti object
}
\description{
In this exercise you have to calculate a confidence interval for a proportion.
}
