% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lage.R
\name{lage}
\alias{lage}
\title{Exercise lage}
\usage{
lage(
  seed = sample.int(10000, 1),
  study = lage_davis(seed = seed),
  question = lage_question()
)
}
\arguments{
\item{seed}{seed which is passed to study function by default. If you use
your own study function, this is ignored, unless you pass it directly to
your function.}

\item{study}{list with elements vector, story and seed, by default
lage_davis() with seed taken from}

\item{question}{question data frame, default is lage_question()}
}
\value{
Entry object of rqti class
}
\description{
lage creates a qti Entry exercise for descriptive statistics
}
\details{
Check out the defaults, which are just functions to generate vector,
story and the question data frame.
}
