% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lage.R
\name{lage_davis}
\alias{lage_davis}
\title{Create davis data set for lage exercise}
\usage{
lage_davis(dv = NULL, n = NULL, seed = sample.int(1000, 1))
}
\arguments{
\item{dv}{"Körpergröße" or "Körpergewicht", which is actually the difference in height and weight, one randomly selected if dv = NULL (default).}

\item{n}{sample size, defaults to random selection of 8, 10, or 12, when n = NULL}

\item{seed}{seed to reproduce data set}
}
\value{
list with vector and story
}
\description{
Create davis data set for lage exercise
}
\details{
This handles reasonable filtering. Note that it makes sense to use an even
n because this increases difficulty. At least for the formulas we use (Tiefe_Median = (n+1)/2 and Tiefe_Quartil = (|\emph{Tiefe_Median}|+1)/2).

Further note that the dv (Körpergröße and Körpergewicht) is actually the difference between reported and actual weight and height. I find it more interesting to see this difference as it reveals biases in human perception.
}
