% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lage.R
\name{lage_klausur}
\alias{lage_klausur}
\title{Exercise lage (exam version)}
\usage{
lage_klausur(seeds, question = lage_question()[c(1:3, sample(c(4, 5, 9), 1))])
}
\arguments{
\item{seeds}{seeds for exercises}

\item{question}{selection from \code{lage_question()}}
}
\value{
list of Entry rqti objects
}
\description{
This is a helper function to create a list of lage exercises for exams.
}
\details{
This is useful to create parallel versions of the exercise for an
exam. It always takes the same questions, but varies other aspects of the
exercise. Long calculations are avoided by not taking variance, sd, mad.
Currently there is only one lage data creation function, so this is
taken (lage_davis).
}
