% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lage.R
\name{lage_qdf}
\alias{lage_qdf}
\title{Create questions for exercise lage}
\usage{
lage_qdf(x)
}
\arguments{
\item{x}{This is the variable to create questions for.}
}
\value{
data frame containing all information for questions for qti
}
\description{
Create questions for exercise lage
}
