% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lage.R
\name{lage_question}
\alias{lage_question}
\title{question vector that can be used as input for lage functions}
\usage{
lage_question()
}
\description{
question vector that can be used as input for lage functions
}
