% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lage.R
\name{lage_studis}
\alias{lage_studis}
\title{Exercise lage (student version)}
\usage{
lage_studis()
}
\description{
lage default exercise for students with two sections, each containing 20
different variants, the first focusing on simple calculations, the second on
more complex ones. One of each section is selected randomly. Stat tables and
formulas are provided, as well as the scientific calculator. This is
OPAL-specific, not tested on other qti platforms.
}
