% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagemaße.R
\name{lagemaße2}
\alias{lagemaße2}
\title{Create qti item for lagemaße}
\usage{
lagemaße2(
  dv,
  story,
  question = c("Median", "Unteres Quartil", "Oberes Quartil", "Interquartilsabstand",
    "Mittelwert", "Varianz", "Standardabweichung", "Mittlerer absoluter Abstand",
    "Range"),
  identifier = "lagemasse"
)
}
\arguments{
\item{dv}{dependent variable, numeric vector}

\item{story}{story before questions}

\item{question}{which descriptives should be tested?}

\item{identifier}{identifier for assessment item}
}
\value{
Entry object of qti class
}
\description{
Create qti item for lagemaße
}
\details{
Note that identifiers for specific questions are automatically created.
}
