% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagemaße.R
\name{lagemaße_klausur}
\alias{lagemaße_klausur}
\title{returns list of seed lagemaße exercises}
\usage{
lagemaße_klausur(seed, question = lagemaße_question()[c(1:3, sample(4:5, 1))])
}
\description{
returns list of seed lagemaße exercises
}
\details{
Note that variance is computing extensive, so you might not want to use that
for an exam.
}
