% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagemaße.R
\name{lagemaße_question}
\alias{lagemaße_question}
\title{question vector that can be used as input for lagemaße functions}
\usage{
lagemaße_question()
}
\description{
question vector that can be used as input for lagemaße functions
}
