% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaanalyse.R
\name{metaanalyse}
\alias{metaanalyse}
\title{Exercise metaanalyse}
\usage{
metaanalyse(seeds = sample.int(10000, 1))
}
\arguments{
\item{seeds}{seeds for exercises, can be a single value or a vector for
several versions}
}
\description{
5 Questions, 4 require to transform an ES to r. 5th requires to calculate
average weighted ES.
}
