% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodenguru.R
\name{methodenguru_data}
\alias{methodenguru_data}
\title{Data from opal methodenguru course}
\usage{
methodenguru_data(shortName, path = "data-raw/results/")
}
\arguments{
\item{shortName}{Short name of exercise.}

\item{path}{where to download to}
}
\description{
Data from opal methodenguru course
}
