% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodenlehre-package.R
\docType{package}
\name{methodenlehre-package}
\alias{methodenlehre}
\alias{methodenlehre-package}
\title{methodenlehre: Creates exercises and exams for Methodenlehre at TUC and methoden.guru}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
