% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodenlehre.R
\name{methodenlehreI_data}
\alias{methodenlehreI_data}
\title{Data from opal methodenlehreI course}
\usage{
methodenlehreI_data(shortName, path = "data-raw/results/")
}
\arguments{
\item{shortName}{Short name of exercise.}

\item{path}{where to download to}
}
\description{
Data from opal methodenlehreI course
}
