% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodenlehre.R
\name{methodenlehreII_data}
\alias{methodenlehreII_data}
\title{Data from opal methodenlehreII course}
\usage{
methodenlehreII_data(shortName, path = "data-raw/results/")
}
\arguments{
\item{shortName}{Short name of exercise.}

\item{path}{where to download to}
}
\description{
Data from opal methodenlehreII course
}
