% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mml_eq}
\alias{mml_eq}
\title{transform r expression to mml with solution}
\usage{
mml_eq(orig_expr, return_result = F, flags = list(quote(round(3L))), round = 3)
}
\description{
transform r expression to mml with solution
}
