% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power2.R
\name{power2}
\alias{power2}
\title{Power2 exercise}
\usage{
power2(case = sample(power2_cases(), 1))
}
\arguments{
\item{case}{which case to construct, see \code{power2_cases()}}
}
\value{
OneInRowTable
}
\description{
In this single-choice table exercise, power2, you are presented with six
different scenarios related to a significance test setup. Your task is to
identify the impact on specific parameters when one of them is altered.
}
