% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_type.R
\name{power_type}
\alias{power_type}
\title{Exercise power type}
\usage{
power_type(
  seed = sample.int(10000, 1),
  type = c("a-priori", "post-hoc", "Kompromiss")
)
}
\arguments{
\item{seed}{seed for exercise, defaults to a randomly drawn 4-digits seed}

\item{type}{Which type of power analysis should be selected, default is all}
}
\description{
Different types of power analyses are described and the participant has to decide which parameters must be specified.
}
