% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{prepare_data_exercises}
\alias{prepare_data_exercises}
\title{exercise level, treat each exercise as unit, e.g. can be used for
skalenniveaus or for exams?}
\usage{
prepare_data_exercises(d)
}
\description{
exercise level, treat each exercise as unit, e.g. can be used for
skalenniveaus or for exams?
}
