% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{prepare_items}
\alias{prepare_items}
\title{Prepares data for subsequent analysis. This is for independent exercises, not
exams. Summarizes overall score for the exercise, which allows to calculate
r_it on item level.}
\usage{
prepare_items(d, percent_complete = 0.5)
}
\description{
Prepares data for subsequent analysis. This is for independent exercises, not
exams. Summarizes overall score for the exercise, which allows to calculate
r_it on item level.
}
