% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{regression}
\alias{regression}
\title{Exercise regression}
\usage{
regression(
  seeds = sample.int(1000000, 1),
  study = sample(prep_functions(), 1),
  which_questions = 1:11
)
}
\arguments{
\item{study}{Study scenario, by default a random one is drawn. Check out the function \code{pref_functions()} for all studies.}

\item{which_questions}{integer vector for question selection}

\item{seed}{seed for randomization}
}
\value{
rqti Entry object
}
\description{
In this exercise the student gets a regression table and graphic and must
answer 11 questions.
}
