% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{regression_example}
\alias{regression_example}
\title{takes prepare function and produces a list with all variables needed
to create regression exercise}
\usage{
regression_example(prepare_function, seed)
}
\description{
takes prepare function and produces a list with all variables needed
to create regression exercise
}
