% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_multiple.R
\name{regression_multiple}
\alias{regression_multiple}
\title{Exercise regression_multiple}
\usage{
regression_multiple(seeds = sample.int(10000, 1))
}
\arguments{
\item{seeds}{seeds for exercises, can be a single value or a vector for
several exercises, defaults to randomly drawn 4-digits seeds.}
}
\description{
Multiple regression setting with prestige data. 3 questions: what is the best
predictor? What is the standardized regression coefficient? Estimate of dv
for specific values of iv
}
