% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{regression_stud}
\alias{regression_stud}
\title{Exercise regression (student versions)}
\usage{
regression_stud(seeds = 1:20)
}
\arguments{
\item{seeds}{numeric vector of seeds}
}
\value{
rqti Test object
}
\description{
Creates several versions of \code{regression}, puts them into a section to select
one at random and puts the section into a test.
}
