% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagemaße.R
\name{s1}
\alias{s1}
\title{generate first section exercises, focusing on simple measures}
\usage{
s1(seed)
}
\value{
qti object of class Entry with 3 gaps
}
\description{
generate first section exercises, focusing on simple measures
}
