% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagemaße.R
\name{s2}
\alias{s2}
\title{generate second section exercises, focusing on more complex measures.}
\usage{
s2(seed)
}
\value{
qti object of class entry with 2 gaps
}
\description{
generate second section exercises, focusing on more complex measures.
}
