% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skalenniveau.R
\name{skalenniveau}
\alias{skalenniveau}
\title{Create question for exercise skalenniveau}
\usage{
skalenniveau(n = 2)
}
\arguments{
\item{n}{number of questions}
}
\value{
OneInRowTable item object
}
\description{
Draws a random sample of n exercises from the dataset skalenniveau from this
package. This dataset currently contains 56 questions. You can also just load
this dataset and select the questions yourself. Use df2sc2 to translate it to
a single choice table.
}
