% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skalenniveau.R
\name{skalenniveau_klausur}
\alias{skalenniveau_klausur}
\title{Create question for exercise skalenniveau for exam}
\usage{
skalenniveau_klausur(n = 2, seed = sample.int(10000, 1))
}
\description{
Difference to skalenniveau function is that we select only diagnostically
useful questions from the 56 available. Currently this amount to 12 questions
}
