% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skalenniveau.R
\name{skalenniveau_studis}
\alias{skalenniveau_studis}
\title{Exercise Skalenniveau}
\usage{
skalenniveau_studis(sample_n = 5)
}
\arguments{
\item{n}{number of items to draw for one run}
}
\value{
test object AssessmentTestOpal
}
\description{
Create skalenniveau qti exercise for students.
}
\details{
One section with 5 elements randomly drawn.
}
