% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{ttest}
\alias{ttest}
\title{Exercise ttest}
\usage{
ttest(seeds = sample.int(10000, 1))
}
\arguments{
\item{seeds}{Seeds for exercises, can be a single value or a vector}
}
\description{
An independent t-test result based on real data is presesented. 4 questions
have to be answered.
}
