% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity.R
\name{validity}
\alias{validity}
\title{Exercise validity}
\usage{
validity(scenario = sample(validity_scenarios(), 1))
}
\arguments{
\item{scenario}{The scenario to choose from, see \code{validity_scenarios} for
options}
}
\description{
Currently only the Dutton & Aron 1974 study is included with different
scenarios. 3 questions regarding validity and design must be answered.
}
