% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{webshot}
\alias{webshot}
\title{take screenshot of opal website with capture-website-cli
requires capture-website-cli on system!}
\usage{
webshot(url, file, width = 800, height = 2000)
}
\description{
take screenshot of opal website with capture-website-cli
requires capture-website-cli on system!
}
