% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wissenschaftstheorie.R
\name{wissenschaftstheorie}
\alias{wissenschaftstheorie}
\title{Exercise wissenschaftstheorie}
\usage{
wissenschaftstheorie(seed = sample.int(1000, 1))
}
\arguments{
\item{seed}{seed for exercise}
}
\description{
5 statements are randomly drawn from 13, student has to assign the correct
philosophy of sicence.
}
