% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zstand.R
\name{zstand}
\alias{zstand}
\title{Exercise zstand}
\usage{
zstand(seed = sample.int(10000, 1))
}
\arguments{
\item{seed}{See for the exercise.}
}
\value{
Entry rqti object
}
\description{
In this exercise the candidate has to calculate the raw value of a z-value and also find out the corresponding p-value for the z-value.
}
\examples{
z <- zstand()
}
