% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova}
\alias{anova}
\title{anova exercise}
\usage{
anova(seed = sample.int(1e+05, 1))
}
\arguments{
\item{seed}{Seed for the exercise. Default is a random seed drawn from 1:1e5.}
}
\value{
rqti Entry object
}
\description{
A realistic within ANOVA output is presented based on a real study. 8
questions have to be answered. This is the raw version, producing an rqti
Entry object.
}
\examples{
a <- anova()
}
