% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagemaße.R
\name{lagemaße}
\alias{lagemaße}
\title{lagemaße creates a qti Entry exercise for descriptive statistics}
\usage{
lagemaße(study = lagemaße_davis(), question = lagemaße_question())
}
\value{
Entry object of qti class
}
\description{
lagemaße creates a qti Entry exercise for descriptive statistics
}
\details{
Check out the defaults, which are just functions to generate vector, story
and the question data frame.
}
