% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power2.R
\name{power2}
\alias{power2}
\title{Power2 exercise}
\usage{
power2(case = sample(power2_cases(), 1))
}
\arguments{
\item{case}{which case to construct, see get_data(power2_data), column id, default
is one random case}
}
\value{
OneInRowTable
}
\description{
identify the impact on specific parameters when one of them is altered.
}
