% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stattests.R
\name{stattest}
\alias{stattest}
\title{stattests exercise}
\usage{
stattest(n_questions = 1)
}
\arguments{
\item{n_questions}{number of questions to create}
}
\value{
list of exercises
}
\description{
The exercise consists of selecting the appropriate test for a study. There
are 14 variants and 12 different tests.
}
